# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * АЛИ-БАБА И 40 КЛАДОВ — Multiplayer board game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginCredentials {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  player: 'player',
} as const;

export interface User {
  id: number;
  username: string;
  role: UserRole;
}

export interface AuthResponse {
  user: User;
}

export type RoomStatus = typeof RoomStatus[keyof typeof RoomStatus];


export const RoomStatus = {
  waiting: 'waiting',
  active: 'active',
  finished: 'finished',
} as const;

export interface Room {
  id: number;
  name: string;
  status: RoomStatus;
  createdAt: string;
}

export interface RoomInput {
  /** @minLength 1 */
  name: string;
}

export interface GamePlayer {
  id: number;
  roomId: number;
  userId: number;
  username: string;
  /**
     * @minimum 0
     * @maximum 60
     */
  position: number;
  color: string;
  canMove: boolean;
}

export type RoomDetailStatus = typeof RoomDetailStatus[keyof typeof RoomDetailStatus];


export const RoomDetailStatus = {
  waiting: 'waiting',
  active: 'active',
  finished: 'finished',
} as const;

export interface RoomDetail {
  id: number;
  name: string;
  status: RoomDetailStatus;
  createdAt: string;
  players: GamePlayer[];
}