# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: АЛИ-БАБА И 40 КЛАДОВ — Multiplayer board game API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: rooms
    description: Game rooms
  - name: players
    description: Players in rooms
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginCredentials"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout current user
      responses:
        "200":
          description: Logged out

  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List all game rooms
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a new game room (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "403":
          description: Forbidden

  /rooms/{id}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get a room by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Room details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomDetail"
        "404":
          description: Not found

  /rooms/{id}/players:
    get:
      operationId: getRoomPlayers
      tags: [players]
      summary: Get all players in a room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GamePlayer"

  /rooms/{id}/join:
    post:
      operationId: joinRoom
      tags: [rooms]
      summary: Join a game room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GamePlayer"
        "400":
          description: Already in room or room full

  /rooms/{id}/start:
    post:
      operationId: startGame
      tags: [rooms]
      summary: Start the game (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Game started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "403":
          description: Forbidden

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    LoginCredentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1

    User:
      type: object
      required: [id, username, role]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, player]

    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"

    Room:
      type: object
      required: [id, name, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [waiting, active, finished]
        createdAt:
          type: string
          format: date-time

    RoomInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1

    GamePlayer:
      type: object
      required: [id, roomId, userId, username, position, color, canMove]
      properties:
        id:
          type: integer
        roomId:
          type: integer
        userId:
          type: integer
        username:
          type: string
        position:
          type: integer
          minimum: 0
          maximum: 60
        color:
          type: string
        canMove:
          type: boolean

    RoomDetail:
      type: object
      required: [id, name, status, createdAt, players]
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [waiting, active, finished]
        createdAt:
          type: string
          format: date-time
        players:
          type: array
          items:
            $ref: "#/components/schemas/GamePlayer"