# lib/api-zod/src/generated/types/gamePlayer.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * АЛИ-БАБА И 40 КЛАДОВ — Multiplayer board game API
 * OpenAPI spec version: 0.1.0
 */

export interface GamePlayer {
  id: number;
  roomId: number;
  userId: number;
  username: string;
  /**
     * @minimum 0
     * @maximum 60
     */
  position: number;
  color: string;
  canMove: boolean;
}