# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * АЛИ-БАБА И 40 КЛАДОВ — Multiplayer board game API
 * OpenAPI spec version: 0.1.0
 */
import type { RoomStatus } from './roomStatus';

export interface Room {
  id: number;
  name: string;
  status: RoomStatus;
  createdAt: Date;
}