# lib/api-zod/src/generated/types/roomDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * АЛИ-БАБА И 40 КЛАДОВ — Multiplayer board game API
 * OpenAPI spec version: 0.1.0
 */
import type { GamePlayer } from './gamePlayer';
import type { RoomDetailStatus } from './roomDetailStatus';

export interface RoomDetail {
  id: number;
  name: string;
  status: RoomDetailStatus;
  createdAt: Date;
  players: GamePlayer[];
}